+++ src/components/selling-item/SellingItem.tsx
@@ -2,10 +2,11 @@ import { useEffect } from 'react';
 import Swal from 'sweetalert2';
 import { useAppDispatch, useAppSelector } from '../../store';
 import { setBlockchainTransactionStatus, spendTokensAsyncAction } from '../../store/actions/blockchain-actions';
-import { deleteMarketProductActionAsync, setMarketProducts, setMarketProductsShowModalAction, setMerketProductSelectedItemAction } from '../../store/actions/market-products-actions';
+import { deleteMarketProductActionAsync, setMarketProducts, setMarketProductsShowModalAction, setMerketProductSelectedItemAction, updateSellingItemActionAsync } from '../../store/actions/market-products-actions';
 import { postNewTransactionActionAsync } from '../../store/actions/transaction-actions';
 import { blockchainAccountSelector, blockchainTransactionStatusSelector } from '../../store/selectors/blockchain-selectors';
 import { contractInfoIsAdminSelector } from '../../store/selectors/contract-info-selectors';
+import { marketPlaceHowMuchSelector, marketPlaceSelectedItemSelector } from '../../store/selectors/market-products-selectors';
 import IMarketProduct from '../../types/IMarketProduct';
 import { ButtonsSection, BuyButton, DeleteButton, LateItemData, LatestItemImage, LatestItemName, LatestItemPrice, LatestItemSectionContainer, LatestItemTokenPrice } from './styles'
 type Props = {
@@ -16,9 +17,29 @@ const SellingItem = (props: Props) => {
     const dispatch = useAppDispatch();
     const isAdmin = useAppSelector(contractInfoIsAdminSelector);
     const accountAddress= useAppSelector(blockchainAccountSelector);
+    const howMuch = useAppSelector(marketPlaceHowMuchSelector);
+    const selectedItem = useAppSelector(marketPlaceSelectedItemSelector);
+
     const transactionStatus = useAppSelector(blockchainTransactionStatusSelector);
     const handleDeleteClick = () => {
-        dispatch(deleteMarketProductActionAsync(id));
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+              if (result.isConfirmed) {
+                  dispatch(deleteMarketProductActionAsync(id));
+                  Swal.fire(
+                'Deleted!',
+                'Your product has been deleted.',
+                'success'
+              )
+            }
+          })
     }
     const handleBuyClick = () =>{
         if(!accountAddress){
@@ -33,15 +54,16 @@ const SellingItem = (props: Props) => {
         dispatch(setMarketProductsShowModalAction(true));
     }
     useEffect(() =>{
-        if(transactionStatus === 1){
+        if(transactionStatus === 1 && selectedItem?.id === props.product.id){
         const sendTransaction = async() =>{
 
+        await dispatch(updateSellingItemActionAsync({howMuch, id}))
         
         await dispatch(postNewTransactionActionAsync({
             address: accountAddress,
             amount: price,
             name,
-            howMuch: 3,
+            howMuch,
             createdAt: new Date(Date.now()),
         }));
     }
